+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 This file documents all notable changes to this project.
 
-## [x.x.x] - TBD
+## [0.7.0] - TBD
 ### Changed
-- use strict mypy
+- Support Enums
+- Add default values
+- Use jinja2 templates
+- Use strict mypy
+- Drop support for Python 3.7
 
 ## [0.6.0] - 2024-01-14